refactor: abstract JSON-RPC code
